+++ parser0.h
@@ -58,7 +58,7 @@ public:
 		; //NOOP
 	}
 	char	Parse(int charset, const std::vector<CDic1>& dics);
-	char	ParseEmbedString(yaya::string_t& str, CStatement &st, const yaya::string_t &dicfilename, int linecount);
+	char	ParseEmbedString(yaya::string_t& str, CStatement &st, const yaya::string_t &dicfilename, ptrdiff_t linecount);
 
 	int		DynamicLoadDictionary(const yaya::string_t& dicfilename, int charset);
 	int		DynamicAppendRuntimeDictionary(const yaya::string_t& codes);
@@ -72,38 +72,37 @@ protected:
 	bool	ParseAfterLoad(const yaya::string_t &dicfilename);
 	char	LoadDictionary1(const yaya::string_t& filename, std::vector<CDefine>& gdefines, int charset);
 	char	GetPreProcess(yaya::string_t& str, std::vector<CDefine>& defines, std::vector<CDefine>& gdefines, const yaya::string_t& dicfilename,
-					int linecount);
+			ptrdiff_t linecount);
 
-	void	ExecInternalPreProcess(yaya::string_t &str,const yaya::string_t &file,int line);
+	void	ExecInternalPreProcess(yaya::string_t &str,const yaya::string_t &file, ptrdiff_t line);
 
 	char	IsCipheredDic(const yaya::string_t& filename);
 	void	SeparateFactor(std::vector<yaya::string_t> &s, yaya::string_t &line);
-	char	DefineFunctions(std::vector<yaya::string_t> &s, const yaya::string_t& dicfilename, int linecount, int &depth, int &targetfunction);
-	int		MakeFunction(const yaya::string_t& name, choicetype_t chtype, const yaya::string_t& dicfilename, int linecount);
-	char	StoreInternalStatement(int targetfunc, yaya::string_t &str, int& depth, const yaya::string_t& dicfilename, int linecount);
-	char	MakeStatement(int type, int targetfunc, yaya::string_t &str, const yaya::string_t& dicfilename, int linecount);
-	char	StructWhen(yaya::string_t &str, std::vector<CCell> &cells, const yaya::string_t& dicfilename, int linecount);
-	char	StructFormula(yaya::string_t &str, std::vector<CCell> &cells, const yaya::string_t& dicfilename, int linecount);
+	char	DefineFunctions(std::vector<yaya::string_t> &s, const yaya::string_t& dicfilename, ptrdiff_t linecount, size_t&depth, ptrdiff_t&targetfunction);
+	ptrdiff_t MakeFunction(const yaya::string_t& name, choicetype_t chtype, const yaya::string_t& dicfilename, ptrdiff_t linecount);
+	char	StoreInternalStatement(size_t targetfunc, yaya::string_t& str, size_t& depth, const yaya::string_t& dicfilename, ptrdiff_t linecount);
+	char	MakeStatement(int type, size_t targetfunc, yaya::string_t &str, const yaya::string_t& dicfilename, ptrdiff_t linecount);
+	char	StructWhen(yaya::string_t &str, std::vector<CCell> &cells, const yaya::string_t& dicfilename, ptrdiff_t linecount);
+	char	StructFormula(yaya::string_t &str, std::vector<CCell> &cells, const yaya::string_t& dicfilename, ptrdiff_t linecount);
 	void	StructFormulaCell(yaya::string_t &str, std::vector<CCell> &cells);
 
 	char	AddSimpleIfBrace(const yaya::string_t &dicfilename);
 
 	char	SetCellType(const yaya::string_t &dicfilename);
-	char	SetCellType1(CCell& scell, char emb, const yaya::string_t& dicfilename, int linecount);
+	char	SetCellType1(CCell& scell, char emb, const yaya::string_t& dicfilename, ptrdiff_t linecount);
 
 	char	MakeCompleteFormula(const yaya::string_t &dicfilename);
 	char	ParseEmbeddedFactor(const yaya::string_t& dicfilename);
 	char	ParseEmbeddedFactor1(CStatement& st, const yaya::string_t& dicfilename);
 	void	ConvertPlainString(const yaya::string_t& dicfilename);
 	void	ConvertPlainString1(CStatement& st, const yaya::string_t& dicfilename);
-	char	ConvertEmbedStringToFormula(yaya::string_t& str, const yaya::string_t& dicfilename, int linecount);
+	char	ConvertEmbedStringToFormula(yaya::string_t& str, const yaya::string_t& dicfilename, ptrdiff_t linecount);
 	char	CheckDepthAndSerialize(const yaya::string_t& dicfilename);
 	char	CheckDepth1(CStatement& st, const yaya::string_t& dicfilename);
 	char	CheckDepthAndSerialize1(CStatement& st, const yaya::string_t& dicfilename);
 	char	MakeCompleteConvertionWhenToIf(const yaya::string_t& dicfilename);
 
 	char	IsDicFileAlreadyExist(const yaya::string_t& dicfilename);
-
 };
 
 //----
